+++ src/main/java/com/nimble/server_spring/infra/security/dto/response/LoginResponseDto.java
@@ -1,21 +1,24 @@
-package com.nimble.server_spring.modules.auth.dto.response;
+package com.nimble.server_spring.infra.security.dto.response;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.nimble.server_spring.modules.auth.JwtToken;
-import lombok.AllArgsConstructor;
 import lombok.Builder;
-import lombok.Data;
+import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.SneakyThrows;
 
-@Data
-@AllArgsConstructor
+@Getter
 @NoArgsConstructor
-@Builder
 public class LoginResponseDto {
 
     private Long userId;
     private String accessToken;
+
+    @Builder
+    private LoginResponseDto(Long userId, String accessToken) {
+        this.userId = userId;
+        this.accessToken = accessToken;
+    }
 
     public static LoginResponseDto fromJwtToken(JwtToken jwtToken) {
         return LoginResponseDto.builder()
