+++ src/main/java/com/nimble/server_spring/modules/auth/AuthController.java
@@ -9,9 +9,9 @@ import com.nimble.server_spring.infra.error.ErrorCodeException;
 import com.nimble.server_spring.infra.jwt.JwtProperties;
 import com.nimble.server_spring.infra.http.CookieParser;
 import com.nimble.server_spring.infra.http.BearerTokenParser;
-import com.nimble.server_spring.modules.auth.dto.request.LocalLoginRequestDto;
+import com.nimble.server_spring.infra.security.dto.request.LocalLoginRequestDto;
 import com.nimble.server_spring.modules.auth.dto.request.LocalSignupRequestDto;
-import com.nimble.server_spring.modules.auth.dto.response.LoginResponseDto;
+import com.nimble.server_spring.infra.security.dto.response.LoginResponseDto;
 import com.nimble.server_spring.modules.auth.dto.response.UserResponseDto;
 import com.nimble.server_spring.modules.user.User;
 import com.nimble.server_spring.modules.user.UserService;
